CINDER-72 fixing the the cg test

+++ tests/integration/test_consistencygroups.py
@@ -58,7 +58,7 @@ class CGRealTestCaseMixin(test_case.RealTestCaseMixin):
 
 class CGTestsMixin(object):
     @contextmanager
-    def volume_context(self, name, pool, consistencygroup_id=None):
+    def volume_context(self, name, pool, consistencygroup_id=None, delete=True):
         from cinderclient.v2.consistencygroups import ConsistencygroupManager
         from cinderclient.v2.volumes import VolumeManager
         vm = VolumeManager(get_cinder_v2_client())
@@ -71,12 +71,13 @@ class CGTestsMixin(object):
         try:
             yield vol
         finally:
-            if vol.consistencygroup_id:
-                cg = cgm.get(vol.consistencygroup_id)
-                cg.update(remove_volumes=vol.id)
-                self.wait_for_removal_from_consistencygroup(vol, timeout=30)
-            vol.delete()
-            self.wait_for_object_deletion(vol, timeout=30)
+            if delete:
+                if vol.consistencygroup_id:
+                    cg = cgm.get(vol.consistencygroup_id)
+                    cg.update(remove_volumes=vol.id)
+                    self.wait_for_removal_from_consistencygroup(vol, timeout=30)
+                vol.delete()
+                self.wait_for_object_deletion(vol, timeout=30)
 
     @contextmanager
     def cg_context(self, name, pool):
@@ -124,7 +125,7 @@ class CGTestsMixin(object):
 
         sm = SnapshotManager(get_cinder_v2_client())
         with self.provisioning_pool_context() as pool:
-            with self.volume_context(name="vol1", pool=pool) as vol1:
+            with self.volume_context(name="vol1", pool=pool, delete=False) as vol1: # will be deleted along with the cg
                 with self.cg_context(name="cg1", pool=pool) as cg:
                     cg.update(add_volumes=vol1.id)
                     time.sleep(5) # Just to make sure the volume was added (we don't have "wait_for_object_addition")
